関数重複阻止
